Support arbitrary numbers of threads for memory allocation.
When I originally refactored memory.c to reduce locking, I made the (incorrect) assumption that all threads were managed by OpenBLAS.  The recent Issues we've seen show that really, any caller can make its own threads and call into OpenBLAS; they don't all come from blas_server.  Thus we have to be able to support an arbitrary number of threads that can come in any time.  The original implementation (before my changes) dealt with this by having a single allocation table, and everyone had to lock to get access to and update it, which was expensive.  Moving to thread-local allocation tables was much faster, but now we have to deal with the fact that thread local storage might not be cleaned up.

This change gives each thread its own local allocation table, and completely does away with the global table.  We cleanup allocations using pthreads' key destructor and Win32's DLL_THREAD_DETACH.

This change also removes compiler TLS, which in the end, wasn't really worth it given the issues with the glibc implementation.  The overall performance impact was < 1%, anyway.  Removing it also simplifies the code.

Support arbitrary numbers of threads for memory allocation.

When I originally refactored memory.c to reduce locking, I made the (incorrect) assumption that all threads were managed by OpenBLAS.  The recent Issues we've seen show that really, any caller can make its own threads and call into OpenBLAS; they don't all come from blas_server.  Thus we have to be able to support an arbitrary number of threads that can come in any time.  The original implementation (before my changes) dealt with this by having a single allocation table, and everyone had to lock to get access to and update it, which was expensive.  Moving to thread-local allocation tables was much faster, but now we have to deal with the fact that thread local storage might not be cleaned up.

This change gives each thread its own local allocation table, and completely does away with the global table.  We cleanup allocations using pthreads' key destructor and Win32's DLL_THREAD_DETACH.

This change also removes compiler TLS, which in the end, wasn't really worth it given the issues with the glibc implementation.  The overall performance impact was < 1%, anyway.  Removing it also simplifies the code.