Do not attempt to check host CPU if TARGET is set.
This wired 'autodetection' breaks DYNAMIC arch (or makes the build
unreproducible) and will most likely not work with cross compiling.

Presently, this is only relevant for SkylakeX, Cooperlake and
Sapphire Rapids cores, that's what has been implemented. Going
forward, other cores would have to be added here as well (in sync
with Markfile.$(arch) as this needs to be kept in sync with changes
to getarch.
It would be better to remove this hack (and the counterpart in
getarch) entirely.
This fixes issue #3557.

Signed-off-by: Egbert Eich <eich@suse.com>