Always propagate HAVE_/NO_ values from ARCHCONFIG when FORCE is enabled
You can end up in a situation where the HAVE_x/NO_x flags are used for compiling param.h but not
Makefile.conf which leads to further inconsistencies. This uses the existing logic to parse ARCHCONFIG and ensures it ends up
in the eventual Makefile.conf whenever FORCE is enabled.